Changed strategy implementations to be Components
