+++ src/main/java/com/team/ghana/enums/Endpoint.java
@@ -1,0 +1,5 @@
+package com.team.ghana.enums;
+
+public enum Endpoint {
+    BUSINESSUNIT, DEPARTMENT, UNIT
+}
