+++ src/main/java/com/team/ghana/searchEmployeeStrategy/SearchEmployeesByDepartmentStrategy.java
@@ -1,19 +1,18 @@
-package com.team.ghana.strategy;
+package com.team.ghana.searchEmployeeStrategy;
 
 import com.team.ghana.department.DepartmentRepository;
 import com.team.ghana.employee.Employee;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Component;
 
 import java.util.ArrayList;
 import java.util.List;
 
+@Component
 public class SearchEmployeesByDepartmentStrategy implements SearchEmployeeStrategy {
 
+    @Autowired
     private DepartmentRepository departmentRepository;
-
-    public SearchEmployeesByDepartmentStrategy(DepartmentRepository departmentRepository) {
-        this.departmentRepository = departmentRepository;
-    }
 
     @Override
     public List<Employee> execute(List<Employee> allEmployees, Long id) {
