+++ src/main/java/com/team/ghana/searchEmployeeStrategy/SearchEmployeesByUnitStrategy.java
@@ -1,19 +1,18 @@
-package com.team.ghana.strategy;
+package com.team.ghana.searchEmployeeStrategy;
 
 import com.team.ghana.employee.Employee;
 import com.team.ghana.unit.UnitRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Component;
 
 import java.util.ArrayList;
 import java.util.List;
 
+@Component
 public class SearchEmployeesByUnitStrategy implements SearchEmployeeStrategy {
 
+    @Autowired
     private UnitRepository unitRepository;
-
-    public SearchEmployeesByUnitStrategy(UnitRepository unitRepository) {
-        this.unitRepository = unitRepository;
-    }
 
     @Override
     public List<Employee> execute(List<Employee> allEmployees, Long id) {
